Small fix for syncing radio frequencies
Adding frequencies to be synced on initial connect and when changed. For COM1 as well as NAV1. Active and standby frequencies.

If you use SPAD.Next make sure to chnage the evnts for sending frequency changes to COM_RADIO_WHOLE_DEC/INC and COM_RADIO_FRACT_DEC/INC

The initial sync might need to be checked as it didin't really sync when I tried it.