feat: added lights yaml to the DV20
After testing I found out the landing, strobe and nav lights are not synchronized in the DV20. I assume added the lights.yaml to the configuration will solve this. I have no clue how to test this this fork so this is only an assumption. Should be verified with the code owner.